Add the BSDCan 2025 "coin" from my talk
This is the coin I printed during my talk and had available to give
away.

It's not really useful for anyone and I failed at printing it a fair
number of times when printing a full bed of them, but I guess if you
want to pretend you were at BSDCan and got one, you can.